aiservice/requirements.txt
flask
python-dotenv
groq
flask-limiter
pytest
gunicorn